Massive refactorng to reduce 256-bit precision to 128-bit on all token and liquidity flows. Replace FullMath with in-256 bit wherver possible

+++ contracts/CrocSwapBooks.sol
@@ -26,7 +26,7 @@ contract CrocSwapBooks is CurveTrader {
     }
 
     function runInit (PoolSpecs.PoolCursor memory pool, uint128 price)
-        masterOnly public returns (int256 baseFlow, int256 quoteFlow) {
+        masterOnly public returns (int128 baseFlow, int128 quoteFlow) {
         return initCurve(pool, price, initLock_, msg.sender);
     }
