+++ contracts/test/TestCurveMath.sol
@@ -17,9 +17,9 @@ contract TestCurveMath {
             buildCurve(seed, growth, concentrated, 0));
     }
 
-    function testVig (uint128 liq, uint256 swapQty, uint24 feeRate, uint8 protoCut,
+    function testVig (uint128 liq, uint128 swapQty, uint24 feeRate, uint8 protoCut,
                       bool isBuy, bool inBase, uint128 curvePrice, uint128 limitPrice)
-        public pure returns (uint256, uint256) {
+        public pure returns (uint128, uint128) {
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, curvePrice);
         CurveMath.SwapFrame memory cntx = CurveMath.SwapFrame(isBuy, inBase, feeRate, protoCut);
         CurveMath.SwapAccum memory swap = CurveMath.SwapAccum(swapQty, 0, 0, 0, cntx);
@@ -28,8 +28,8 @@ contract TestCurveMath {
 
     function testVigMin (uint128 liq, uint24 feeRate, uint8 protoCut,
                          bool inBase, uint128 curvePrice)
-        public pure returns (uint256, uint256) {
-        uint swapQty = type(uint128).max;
+        public pure returns (uint128, uint128) {
+        uint128 swapQty = type(uint128).max;
         bool isBuy = inBase ? false : true;
         return testVig(liq, swapQty, feeRate, protoCut, isBuy, inBase, curvePrice,
                        TickMath.MIN_SQRT_RATIO);
@@ -37,101 +37,101 @@ contract TestCurveMath {
 
     function testVigMax (uint128 liq, uint24 feeRate, uint8 protoCut,
                          bool inBase, uint128 curvePrice)
-        public pure returns (uint256, uint256) {
-        uint swapQty = type(uint128).max;
+        public pure returns (uint128, uint128) {
+        uint128 swapQty = type(uint128).max;
         bool isBuy = inBase ? true : false;
         return testVig(liq, swapQty, feeRate, protoCut, isBuy, inBase, curvePrice,
                        TickMath.MAX_SQRT_RATIO);
     }
 
     function testLimitBase (uint128 price, uint128 limitPrice, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, price);
         CurveMath.SwapAccum memory swap = buildSwap(1000000, true, true);
         return CurveMath.calcLimitFlows(curve, swap, limitPrice);
     }
 
     function testLimitQuote (uint128 price, uint128 limitPrice, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, price);
         CurveMath.SwapAccum memory swap = buildSwap(1000000, true, false);
         return CurveMath.calcLimitFlows(curve, swap, limitPrice);
     }
 
     function testCounterBase (uint128 price, uint128 limitPrice, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, price);
         CurveMath.SwapAccum memory swap = buildSwap(1000000, true, true);        
         return CurveMath.calcLimitCounter(curve, swap, limitPrice);
     }
 
     function testCounterQuote (uint128 price, uint128 limitPrice, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, price);
         CurveMath.SwapAccum memory swap = buildSwap(1000000, true, false); 
         return CurveMath.calcLimitCounter(curve, swap, limitPrice);
     }
 
     function testLimitBaseMax (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testLimitBase(price, TickMath.MAX_SQRT_RATIO, liq);
     }    
 
     function testLimitBaseMin (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testLimitBase(price, TickMath.MIN_SQRT_RATIO, liq);
     }    
 
     function testLimitQuoteMax (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testLimitQuote(price, TickMath.MAX_SQRT_RATIO, liq);
     }    
 
     function testLimitQuoteMin (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testLimitQuote(price, TickMath.MIN_SQRT_RATIO, liq);
     }    
 
     function testCounterBaseMax (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testCounterBase(price, TickMath.MAX_SQRT_RATIO, liq);
     }    
 
     function testCounterBaseMin (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testCounterBase(price, TickMath.MIN_SQRT_RATIO, liq);
     }    
 
     function testCounterQuoteMax (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testCounterQuote(price, TickMath.MAX_SQRT_RATIO, liq);
     }    
 
     function testCounterQuoteMin (uint128 price, uint128 liq)
-        public pure returns (uint256) {
+        public pure returns (uint128) {
         return testCounterQuote(price, TickMath.MIN_SQRT_RATIO, liq);
     }    
 
     function testLimitQtyLeft (uint128 price, uint128 limitPrice, uint128 liq,
-                               uint256 swapQty)
-        public pure returns (uint256) {
+                               uint128 swapQty)
+        public pure returns (uint128) {
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, price);
         CurveMath.SwapAccum memory swap = buildSwap(swapQty, true, true);
         return CurveMath.calcLimitFlows(curve, swap, limitPrice);
     }
 
     function testCounterQtyLeft (uint128 price, uint128 limitPrice, uint128 liq,
-                                 uint256 swapQty)
-        public pure returns (uint256) {
+                                 uint128 swapQty)
+        public pure returns (uint128) {
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, price);
         CurveMath.SwapAccum memory swap = buildSwap(swapQty, true, true);
         return CurveMath.calcLimitCounter(curve, swap, limitPrice);
     }
 
-    function testRoll (uint256 flow, uint128 price, uint128 liq,
+    function testRoll (uint128 flow, uint128 price, uint128 liq,
                        bool isBuy, bool inBase)
-        public pure returns (uint128 rollPrice, uint256 qtyLeft,
-                             int256 paidBase, int256 paidQuote) {
+        public pure returns (uint128 rollPrice, uint128 qtyLeft,
+                             int128 paidBase, int128 paidQuote) {
         CurveMath.SwapAccum memory swap = buildSwap(flow, isBuy, inBase);
         CurveMath.CurveState memory curve = buildCurve(liq, 0, 0, price);
         CurveRoll.rollFlow(curve, flow, swap);
@@ -140,8 +140,8 @@ contract TestCurveMath {
     }
 
     function testRollInf (uint128 liq, uint128 price, bool isBuy, bool inBase)
-        public pure returns (uint128 rollPrice, uint256 qtyLeft,
-                             int256 paidBase, int256 paidQuote) {
+        public pure returns (uint128 rollPrice, uint128 qtyLeft,
+                             int128 paidBase, int128 paidQuote) {
         uint128 flow = (isBuy == inBase) ? type(uint128).max :
             SafeCast.toUint128(inBase ?
                     FullMath.mulDiv(liq, price, FixedPoint.Q96) :
@@ -150,10 +150,10 @@ contract TestCurveMath {
             testRoll(flow, price, liq, isBuy, inBase);
     }
 
-    function testAssimilate (uint256 feesPaid, uint128 price,
+    function testAssimilate (uint128 feesPaid, uint128 price,
                              uint128 seed, uint128 conc, uint64 growth, bool inBase)
         public pure returns (uint128 shiftPrice, uint128 shiftSeed,
-                             uint256 shiftGrowth, uint256 concGrowth) {
+                             uint128 shiftGrowth, uint128 concGrowth) {
         CurveMath.CurveState memory curve = buildCurve(seed, growth, conc, price);
         CurveAssimilate.assimilateLiq(curve, feesPaid, inBase);
         
@@ -163,9 +163,9 @@ contract TestCurveMath {
     }
 
     function testDeriveImpact (uint128 price, uint128 seed, uint64 growth,
-                               uint128 conc, uint256 flow, 
+                               uint128 conc, uint128 flow, 
                                bool isBuy, bool inBase)
-        public pure returns (uint256, uint128) {
+        public pure returns (uint128, uint128) {
         CurveMath.CurveState memory curve = buildCurve(seed, growth, conc, price);
         CurveMath.SwapAccum memory swap = buildSwap(flow, isBuy, inBase);
         return CurveRoll.deriveImpact(curve, flow, swap.cntx_);
@@ -184,7 +184,7 @@ contract TestCurveMath {
         return cntx;
     }
 
-    function buildSwap (uint256 flow, bool isBuy, bool inBase)
+    function buildSwap (uint128 flow, bool isBuy, bool inBase)
         private pure returns (CurveMath.SwapAccum memory) {
         CurveMath.SwapFrame memory cntx = buildSwapFrame(isBuy, inBase);
         return CurveMath.SwapAccum(flow, 0, 0, 0, cntx);
